--- mcos/optimizer/__init__.py
@@ -1,2 +1,0 @@
-from mcos.optimizer.base import AbstractOptimizer
-from mcos.optimizer.markowitz import MarkowitzOptimizer
